Improved dialect structure
