+++ src/client/components/Editors/Dialect/Sections/Dialect.tsx
@@ -1,11 +1,10 @@
 import * as React from 'react'
 import Box from '@mui/material/Box'
 import InputField from '../../../Parts/Fields/Input'
-import YesNoField from '../../../Parts/Fields/YesNo'
 import MultilineField from '../../../Parts/Fields/Multiline'
+import SelectField from '../../../Parts/Fields/Select'
 import EditorSection from '../../Base/Section'
 import Columns from '../../../Parts/Grids/Columns'
-import * as settings from '../../../../settings'
 import { useStore } from '../store'
 import validator from 'validator'
 
@@ -17,14 +16,10 @@ export default function General() {
         <Box>
           <Title />
           <Description />
-          <CommentChar />
-          <CommentRows />
         </Box>
         <Box>
-          <Header />
-          <HeaderRows />
-          <HeaderJoin />
-          <HeaderCase />
+          <Type />
+          <Format />
         </Box>
       </Columns>
     </EditorSection>
@@ -68,94 +63,32 @@ function Description() {
   )
 }
 
-function CommentChar() {
-  const commentChar = useStore((state) => state.descriptor.commentChar)
+function Type() {
+  const type = useStore((state) => state.type)
   const updateHelp = useStore((state) => state.updateHelp)
-  const updateDescriptor = useStore((state) => state.updateDescriptor)
+  const updateState = useStore((state) => state.updateState)
   return (
-    <InputField
-      label="Comment Char"
-      value={commentChar || settings.DEFAULT_COMMENT_CHAR}
-      onFocus={() => updateHelp('dialect/commentChar')}
-      onChange={(value) => updateDescriptor({ commentChar: value || undefined })}
+    <SelectField
+      label="Type"
+      value={type || ''}
+      options={['file', 'text', 'json', 'table']}
+      onFocus={() => updateHelp('dialect/type')}
+      onChange={(value) => updateState({ type: value || 'table' })}
     />
   )
 }
 
-function CommentRows() {
-  const commentRows = useStore((state) => state.descriptor.commentRows)
+function Format() {
+  const format = useStore((state) => state.format)
   const updateHelp = useStore((state) => state.updateHelp)
-  const updateDescriptor = useStore((state) => state.updateDescriptor)
+  const updateState = useStore((state) => state.updateState)
   return (
-    <InputField
-      label="Comment Rows"
-      value={(commentRows || []).join(',')}
-      onFocus={() => updateHelp('dialect/commentRows')}
-      onChange={(value) =>
-        updateDescriptor({
-          commentRows: value ? value.split(',').map(parseInt) : undefined,
-        })
-      }
-    />
-  )
-}
-
-function Header() {
-  const header = useStore((state) => state.descriptor.header)
-  const updateHelp = useStore((state) => state.updateHelp)
-  const updateDescriptor = useStore((state) => state.updateDescriptor)
-  return (
-    <YesNoField
-      label="Header"
-      value={header ?? settings.DEFAULT_HEADER}
-      onFocus={() => updateHelp('dialect/header')}
-      onChange={(value) => updateDescriptor({ header: value || undefined })}
-    />
-  )
-}
-
-function HeaderRows() {
-  const headerRows = useStore((state) => state.descriptor.headerRows)
-  const updateHelp = useStore((state) => state.updateHelp)
-  const updateDescriptor = useStore((state) => state.updateDescriptor)
-  return (
-    <InputField
-      label="Header Rows"
-      value={headerRows}
-      onFocus={() => updateHelp('dialect/headerRows')}
-      onChange={(headerRows) =>
-        updateDescriptor({
-          headerRows: headerRows ? headerRows.split(',').map(parseInt) : undefined,
-        })
-      }
-    />
-  )
-}
-
-function HeaderJoin() {
-  const headerJoin = useStore((state) => state.descriptor.headerJoin)
-  const updateHelp = useStore((state) => state.updateHelp)
-  const updateDescriptor = useStore((state) => state.updateDescriptor)
-  return (
-    <InputField
-      label="Header Join"
-      value={headerJoin}
-      onFocus={() => updateHelp('dialect/headerJoin')}
-      onChange={(value) => updateDescriptor({ headerJoin: value || undefined })}
-    />
-  )
-}
-
-function HeaderCase() {
-  const headerCase = useStore((state) => state.descriptor.headerCase)
-  const updateHelp = useStore((state) => state.updateHelp)
-  const updateDescriptor = useStore((state) => state.updateDescriptor)
-  return (
-    <YesNoField
-      label="Header Case"
-      value={headerCase ?? settings.DEFAULT_HEADER_CASE}
-      onFocus={() => updateHelp('dialect/headerCase')}
-      onChange={(value) => updateDescriptor({ headerCase: value || undefined })}
+    <SelectField
+      label="Format"
+      value={format || ''}
+      options={['csv', 'excel', 'json']}
+      onFocus={() => updateHelp('dialect/format')}
+      onChange={(value) => updateState({ format: value || 'table' })}
     />
   )
 }
