+++ src/client/components/Editors/Dialect/Sections/Type/index.tsx
@@ -1,0 +1,14 @@
+import * as React from 'react'
+import EmptyFormat from './Empty'
+import TableFormat from './Table'
+import { useStore } from '../../store'
+
+export default function Format() {
+  const type = useStore((state) => state.type)
+  switch (type) {
+    case 'table':
+      return <TableFormat />
+    default:
+      return <EmptyFormat />
+  }
+}
