+++ src/client/components/Editors/Dialect/store.ts
@@ -14,7 +14,8 @@ import help from './help.yaml'
 const DEFAULT_HELP_ITEM = helpers.readHelpItem(help, 'dialect')!
 
 interface State {
-  format?: string
+  type: string
+  format: string
   section: string
   descriptor: types.IDialect
   onChange: (dialect: types.IDialect) => void
@@ -47,7 +48,8 @@ interface State {
 export function makeStore(props: DialectProps) {
   return createStore<State>((set, get) => ({
     descriptor: props.dialect || cloneDeep(settings.INITIAL_DIALECT),
-    format: props.format,
+    type: props.type || 'table',
+    format: props.format || 'csv',
     section: 'dialect',
     onChange: props.onChange || noop,
     helpItem: DEFAULT_HELP_ITEM,
